+++ src/Method/SendPaidMedia.php
@@ -46,6 +46,7 @@ final readonly class SendPaidMedia implements MethodInterface
         private ?bool $allowPaidBroadcast = null,
         private ?int $directMessagesTopicId = null,
         private ?SuggestedPostParameters $suggestedPostParameters = null,
+        private ?int $messageThreadId = null,
     ) {}
 
     public function getHttpMethod(): HttpMethod
@@ -72,6 +73,7 @@ final readonly class SendPaidMedia implements MethodInterface
             [
                 'business_connection_id' => $this->businessConnectionId,
                 'chat_id' => $this->chatId,
+                'message_thread_id' => $this->messageThreadId,
                 'direct_messages_topic_id' => $this->directMessagesTopicId,
                 'star_count' => $this->starCount,
                 'media' => $media,